feat(managed-baremetal): add listing page
