+++ packages/manager/apps/dedicated/client/app/managedBaremetal/index.js
@@ -6,11 +6,21 @@ const moduleName = 'managedBaremetalLazyloading';
 
 angular.module(moduleName, ['ui.router', 'oc.lazyLoad']).config(
   /* @ngInject */ ($stateProvider) => {
+    $stateProvider.state('app.managedBaremetals.**', {
+      url: '/managedBaremetal',
+      lazyLoad: ($transition$) => {
+        const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
+        return import('./managed-baremetal.module').then((mod) =>
+          $ocLazyLoad.inject(mod.default || mod),
+        );
+      },
+    });
+
     $stateProvider.state('app.managedBaremetal.**', {
       url: '/configuration/managedBaremetal/:productId',
       lazyLoad: ($transition$) => {
         const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
-        return import('./managed-baremetal.module').then((mod) =>
+        return import('./details/managed-baremetal.module').then((mod) =>
           $ocLazyLoad.inject(mod.default || mod),
         );
       },
