+++ packages/manager/apps/dedicated/client/app/managedBaremetal/translations/Messages_fr_FR.json
@@ -1,0 +1,3 @@
+{
+  "managed_baremetal_title": "Managed Baremetal"
+}
